# server/openapi.yaml
openapi: 3.0.3
info:
  title: WEBonTour API
  description: API for the WEBonTour CMS
  version: 1.0.0
  contact:
    name: Project Team
  license:
    name: Project License
servers:
  - url: /api
    description: API base path
...
paths:
...
  /setup/status:
    get:
      tags:
        - setup
      summary: Check setup status
      description: Check if initial setup is required (no users exist)
      operationId: getSetupStatus
      responses:
        '200':
          description: Setup status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  needsSetup:
                    type: boolean
                    example: true
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /setup/create-admin:
    post:
      tags:
        - setup
      summary: Create admin user
      description: Create the initial admin user (only works when no users exist)
      operationId: createAdminUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - email
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securepassword123
                email:
                  type: string
                  format: email
                  example: admin@example.com
      responses:
        '201':
          description: Admin user created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                  message:
                    type: string
                    example: Admin user created successfully
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request or setup already completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          example: user
          enum: [user, writer, moderator, admin]
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message